ISSUE-25: First pass
