+++ src/LostSoul.php
@@ -1,8 +1,9 @@
 <?php
   /**
-   * In our world there are many lost souls in various states of mind. With a few hugs their state or at least their
-   * properties can be changed. Once a lost soul is hugged it will inspire them to hug back which ultmatly will make
-   * the world a better place.
+   * There are many lost souls in various states of mind. With a few hugs their state or at least their
+   * properties can be changed. Once a lost soul is hugged it will inspire them to hug back which ultmatly
+   * will make the world a better place. Or so it is believed:
+   * http://health.usnews.com/health-news/health-wellness/articles/2016-02-03/the-health-benefits-of-hugging
    */
 declare(strict_types=1);
 
@@ -14,7 +15,7 @@ use Psr\Hug\Huggable;
  * Class LostSoul
  * @package DeeZone\Hug
  */
-class LostSoul implements Huggable
+class LostSoul extends Huggers
 {
 
     const POSITIVE = 1;
@@ -23,61 +24,6 @@ class LostSoul implements Huggable
     // http://www.sciencemag.org/news/2011/01/hugs-follow-3-second-rule
     const DURATION_MIN = 1;
     const DURATION_MAX = 6;
-
-    // @var int
-    private $warmAndFuzzy;
-
-    // @var int
-    private $hugged;
-
-    /**
-     * At the beginning of every Lost Soul's existance they start with some general feeling. Funny how some of us are
-     * happier than others even though we all start out the same.
-     */
-    public function __construct()
-    {
-        $this->warmAndFuzzy = random_int(0, 100);
-        $this->hugged = 0;
-    }
-
-    /**
-     * Hugs this object.
-     *
-     * All hugs are mutual. An object that is hugged will in turn hug the other object back by calling hug() on the
-     * first parameter. The number of times hugs are exchanged is determined by the return of keepHugging().
-     *
-     * @param Huggable $soul
-     *   The object (soul) that is hugging this object and will get a hug back in return.
-     *
-     * @return Huggable
-     */
-    public function hug(Huggable $soul): Huggable
-    {
-        // An Exception is thrown as self hugging suggests an error in implimentation. Not a show stopper but should
-        // be addressed.
-        if ($soul === $this) {
-            throw new \Exception('WARNING: You should always love yourself but self hugging is not supported in the ' .
-                'PSR-8 specification. An attept at an object hugging itself has been made.');
-        }
-
-        $durationOfHug = $this->determineDurationOfHug();
-        while ($this->keepHugging($this, $soul, $durationOfHug)) {
-            $hugImpact = $this->hugImpact($this, $soul, $durationOfHug);
-
-            // The power of hugs, everyone is really feeling it
-            $this->warmAndFuzzy += $hugImpact;
-            $soul->warmAndFuzzy += $hugImpact;
-
-            // That was great, lets do it again
-            $this->hug($soul);
-
-            $this->hugged++;
-            $soul->hugged++;
-        }
-
-        // For better or worst the hugging has stopped. Time to let go.
-        return $soul;
-    }
 
     /**
      * Determine if another exchange of hugs is desired.
@@ -88,7 +34,7 @@ class LostSoul implements Huggable
      *
      * @return bool
      */
-    private function keepHugging(Huggable $thisSoul, Huggable $otherSoul, $durationOfHug): bool
+    private function keepHugging(Huggable $thisSoul, Huggable $otherSoul, int $durationOfHug): bool
     {
         // Always start from a positive perspective.
         $keepHugging = true;
@@ -155,38 +101,4 @@ class LostSoul implements Huggable
         return $hugImpact;
     }
 
-    /**
-     * Determine the duration for the hug. Minimum to maximum second hugs, anything more and there's something
-     * else going on.
-     *
-     * http://www.sciencemag.org/news/2011/01/hugs-follow-3-second-rule
-     *
-     * @return int
-     */
-    private function determineDurationOfHug()
-    {
-        $durationOfHug = random_int(self::DURATION_MIN, self::DURATION_MAX);
-
-        return $durationOfHug;
-    }
-
-    /**
-     * How much "Warm And Fuzzy" is the object feeling?
-     *
-     * @return int
-     */
-    public function getWarmAndFuzzy(): int
-    {
-        return $this->warmAndFuzzy;
-    }
-
-    /**
-     * How many times has this soul been hugged?
-     *
-     * @return int
-     */
-    public function getTimesHugged(): int
-    {
-        return $this->hugged;
-    }
 }
